+++ NEWS.rst
@@ -4,6 +4,7 @@ New in 0.6:
 - added functionality to accumulate the results of distributions together with
   the integrated results
 - changed the interfaces of functions that return accumulated results
+- WARNING: removed the ``parameters()`` function in ``hep::projector``
 
 New in 0.5:
 ===========